+++ Application/Config/config_nvm.h
@@ -8,7 +8,7 @@ typedef struct sensor_nvm_S
 {
   config_header_t header;
   uint8_t start;
-  uint16_t log_q_cnt;
+  uint32_t log_q_cnt;
 } config_nvm_t;
 
 void load_config_nvm(void);
@@ -20,7 +20,7 @@ void save_config_nvm(void);
 
 
 config_nvm_t *get_config_nvm(void);
-void nvm_set_log_cnt(uint16_t value);
-uint16_t nvm_get_log_cnt(void);
+void nvm_set_log_cnt(uint32_t value);
+uint32_t nvm_get_log_cnt(void);
 
 #endif
